对三张表对应的实体类进行修改

+++ src/main/java/com/youyou/demo/model/Sku.java
@@ -26,7 +26,6 @@ public class Sku extends BaseEntity{
     private String specs;
     private String code;
     private Long stock;
-    private Integer categoryId;
-    private Integer rootCategoryId;
-
+    private Long categoryId;
+    private Long rootCategoryId;
 }
